demo/src/main/java/com/example/demo/controller/UserController.java
package com.example.demo.controller;




import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.*;

import com.example.demo.models.User;
import com.example.demo.services.UserService;

import java.util.List;

@RestController
@RequestMapping("/users")
public class UserController {

    @Autowired
    private UserService userService;

    @GetMapping
    public List<User> getAllUsers() {
        return userService.getAllUsers();
    }

    @GetMapping("/{userId}")
    public ResponseEntity<User> getUserById(@PathVariable String userId) {
        User user = userService.getUserById(userId);
        return ResponseEntity.ok(user);
    }
    
    @PostMapping
    public ResponseEntity<String> registerUser(@RequestBody User user) {
        userService.registerUser(user.getUserId(), user.getUsername());
        return ResponseEntity.ok("User has been registered successfully.");
    }

    @PutMapping("/{userId}")
    public ResponseEntity<String> updateUserScore(@PathVariable String userId, @RequestBody User user) {
        userService.updateUserScore(userId, user.getScore());
        return ResponseEntity.ok("Score has been updated successfully.");
    }

    @DeleteMapping("/{userId}")
    public ResponseEntity<String> deleteUser(@PathVariable String userId) {
        userService.deleteUser(userId);
        return ResponseEntity.ok("User has been deleted successfully.");
    }
}
